stockpile allocator: keep allocated memory in the user space and reuse it
The stockpile allocator is implemented to reduce the burden of handling fault faults in the kernel.
It keep allocated memory in the user space. When malloc/memalign is called, it tries to return
a candidate matching the requirement to the caller if possible.
It is designed to be lock-free; no mutex is involved.

It can reduce roughly 55% minor page faults, with a trade-off of using 1.7x memory.